+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+* Fixed generating `Event` and `Function` structs where the ABI has a tuple with more than 12 fields.
+
+  The `Event` and `Function` structs in those situations will now render without `Debug` nor `PartialEq` support to allows for Rust to compile properly.
+
 ## [0.10.1](https://github.com/streamingfast/substreams-ethereum/releases/tag/v0.10.1)
 
 * Fixed ABI generation, ambiguous 'as_ref' implementation when using alloy-primitive crate.